# hotstock-be-v1/src/modules/auth/auth.controller.ts
import {
  Controller,
  Post,
  Body,
  Req,
  UseGuards,
  HttpCode,
  HttpStatus,
} from '@nestjs/common';
import {
  ApiTags,
  ApiOperation,
  ApiResponse,
  ApiBearerAuth,
} from '@nestjs/swagger';
import { Throttle } from '@nestjs/throttler';
import { FastifyRequest } from 'fastify';
import { AuthService, AuthResponse, TokensResponse } from './auth.service';
import { LoginDto } from './dto/login.dto';
import { RegisterDto } from './dto/register.dto';
import { RefreshTokenDto } from './dto/refresh-token.dto';
import { ChangePasswordDto } from './dto/change-password.dto';
import { ForgotPasswordDto } from './dto/forgot-password.dto';
import { VerifyOtpDto } from './dto/verify-otp.dto';
import { ResetPasswordDto } from './dto/reset-password.dto';
import { JwtAuthGuard } from '../../common/guards/jwt-auth.guard';
import { CurrentUser } from '../../common/decorators/current-user.decorator';
import { JwtPayload } from './interfaces/jwt-payload.interface';

@ApiTags('Auth')
@Controller('auth')
export class AuthController {
  constructor(private readonly authService: AuthService) {}

  // ─── LOGIN ─────────────────────────────────────────────────────────────────

  @Post('login')
  @HttpCode(HttpStatus.OK)
  @Throttle({ medium: { limit: 10, ttl: 60000 } })
  @ApiOperation({ summary: 'Đăng nhập', description: 'Đăng nhập bằng email và mật khẩu' })
  @ApiResponse({ status: 200, description: 'Đăng nhập thành công' })
  @ApiResponse({ status: 401, description: 'Email hoặc mật khẩu không chính xác' })
  @ApiResponse({ status: 429, description: 'Quá nhiều yêu cầu' })
  async login(
    @Body() dto: LoginDto,
    @Req() request: FastifyRequest,
  ): Promise<AuthResponse> {
    const ip = request.ip;
    const userAgent = request.headers['user-agent'] ?? '';
    return this.authService.login(dto, ip, userAgent);
  }

  // ─── REGISTER ──────────────────────────────────────────────────────────────

  @Post('register')
  @Throttle({ medium: { limit: 10, ttl: 60000 } })
  @ApiOperation({ summary: 'Đăng ký', description: 'Tạo tài khoản mới' })
  @ApiResponse({ status: 201, description: 'Đăng ký thành công' })
  @ApiResponse({ status: 409, description: 'Email đã được sử dụng' })
  @ApiResponse({ status: 429, description: 'Quá nhiều yêu cầu' })
  async register(@Body() dto: RegisterDto): Promise<AuthResponse> {
    return this.authService.register(dto);
  }

  // ─── REFRESH TOKEN ─────────────────────────────────────────────────────────

  @Post('refresh')
  @HttpCode(HttpStatus.OK)
  @ApiOperation({ summary: 'Làm mới token', description: 'Đổi refresh token lấy cặp token mới' })
  @ApiResponse({ status: 200, description: 'Token mới được cấp' })
  @ApiResponse({ status: 401, description: 'Refresh token không hợp lệ' })
  async refresh(
    @Body() dto: RefreshTokenDto,
    @Req() request: FastifyRequest,
  ): Promise<TokensResponse> {
    const ip = request.ip;
    const userAgent = request.headers['user-agent'] ?? '';
    return this.authService.refresh(dto.refresh_token, ip, userAgent);
  }

  // ─── LOGOUT ────────────────────────────────────────────────────────────────

  @Post('logout')
  @HttpCode(HttpStatus.OK)
  @UseGuards(JwtAuthGuard)
  @ApiBearerAuth()
  @ApiOperation({ summary: 'Đăng xuất', description: 'Thu hồi refresh token' })
  @ApiResponse({ status: 200, description: 'Đăng xuất thành công' })
  @ApiResponse({ status: 401, description: 'Chưa đăng nhập' })
  async logout(
    @CurrentUser() user: JwtPayload,
    @Body() dto: RefreshTokenDto,
  ): Promise<{ message: string }> {
    await this.authService.logout(user.sub, dto.refresh_token);
    return { message: 'Đăng xuất thành công' };
  }

  // ─── CHANGE PASSWORD ──────────────────────────────────────────────────────

  @Post('change-password')
  @HttpCode(HttpStatus.OK)
  @UseGuards(JwtAuthGuard)
  @ApiBearerAuth()
  @ApiOperation({ summary: 'Đổi mật khẩu', description: 'Đổi mật khẩu (cần đăng nhập)' })
  @ApiResponse({ status: 200, description: 'Đổi mật khẩu thành công' })
  @ApiResponse({ status: 401, description: 'Mật khẩu cũ không chính xác' })
  async changePassword(
    @CurrentUser() user: JwtPayload,
    @Body() dto: ChangePasswordDto,
  ): Promise<{ message: string }> {
    await this.authService.changePassword(user.sub, dto);
    return { message: 'Đổi mật khẩu thành công' };
  }

  // ─── FORGOT PASSWORD ──────────────────────────────────────────────────────

  @Post('forgot-password')
  @HttpCode(HttpStatus.OK)
  @Throttle({ long: { limit: 3, ttl: 600000 } })
  @ApiOperation({ summary: 'Quên mật khẩu', description: 'Gửi mã OTP đặt lại mật khẩu qua email' })
  @ApiResponse({ status: 200, description: 'Nếu email tồn tại, mã OTP đã được gửi' })
  @ApiResponse({ status: 429, description: 'Quá nhiều yêu cầu' })
  async forgotPassword(
    @Body() dto: ForgotPasswordDto,
  ): Promise<{ message: string }> {
    return this.authService.forgotPassword(dto.email);
  }

  // ─── VERIFY OTP ────────────────────────────────────────────────────────────

  @Post('verify-otp')
  @HttpCode(HttpStatus.OK)
  @Throttle({ long: { limit: 5, ttl: 600000 } })
  @ApiOperation({ summary: 'Xác minh OTP', description: 'Xác minh mã OTP và nhận reset token' })
  @ApiResponse({ status: 200, description: 'OTP hợp lệ, trả về reset token' })
  @ApiResponse({ status: 400, description: 'OTP không hợp lệ hoặc hết hạn' })
  @ApiResponse({ status: 429, description: 'Quá nhiều lần thử' })
  async verifyOtp(
    @Body() dto: VerifyOtpDto,
  ): Promise<{ reset_token: string }> {
    return this.authService.verifyOtp(dto.email, dto.otp);
  }

  // ─── RESET PASSWORD ───────────────────────────────────────────────────────

  @Post('reset-password')
  @HttpCode(HttpStatus.OK)
  @ApiOperation({ summary: 'Đặt lại mật khẩu', description: 'Đặt lại mật khẩu bằng reset token' })
  @ApiResponse({ status: 200, description: 'Đặt lại mật khẩu thành công' })
  @ApiResponse({ status: 401, description: 'Token không hợp lệ hoặc đã hết hạn' })
  async resetPassword(
    @Body() dto: ResetPasswordDto,
  ): Promise<{ message: string }> {
    await this.authService.resetPassword(dto.resetToken, dto.newPassword);
    return { message: 'Đặt lại mật khẩu thành công' };
  }
}