# hotstock-be-v1/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "Đăng nhập bằng email và mật khẩu",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đăng nhập thành công"
          },
          "401": {
            "description": "Email hoặc mật khẩu không chính xác"
          },
          "429": {
            "description": "Quá nhiều yêu cầu"
          }
        },
        "summary": "Đăng nhập",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Tạo tài khoản mới",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đăng ký thành công"
          },
          "409": {
            "description": "Email đã được sử dụng"
          },
          "429": {
            "description": "Quá nhiều yêu cầu"
          }
        },
        "summary": "Đăng ký",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Đổi refresh token lấy cặp token mới",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token mới được cấp"
          },
          "401": {
            "description": "Refresh token không hợp lệ"
          }
        },
        "summary": "Làm mới token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Thu hồi refresh token",
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đăng xuất thành công"
          },
          "401": {
            "description": "Chưa đăng nhập"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Đăng xuất",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "description": "Đổi mật khẩu (cần đăng nhập)",
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đổi mật khẩu thành công"
          },
          "401": {
            "description": "Mật khẩu cũ không chính xác"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Đổi mật khẩu",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "description": "Gửi mã OTP đặt lại mật khẩu qua email",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nếu email tồn tại, mã OTP đã được gửi"
          },
          "429": {
            "description": "Quá nhiều yêu cầu"
          }
        },
        "summary": "Quên mật khẩu",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-otp": {
      "post": {
        "description": "Xác minh mã OTP và nhận reset token",
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP hợp lệ, trả về reset token"
          },
          "400": {
            "description": "OTP không hợp lệ hoặc hết hạn"
          },
          "429": {
            "description": "Quá nhiều lần thử"
          }
        },
        "summary": "Xác minh OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "description": "Đặt lại mật khẩu bằng reset token",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đặt lại mật khẩu thành công"
          },
          "401": {
            "description": "Token không hợp lệ hoặc đã hết hạn"
          }
        },
        "summary": "Đặt lại mật khẩu",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/portfolios/all": {
      "get": {
        "description": "Trả về tất cả danh mục đầu tư để quản lý trong Admin.",
        "operationId": "PortfoliosController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy tất cả danh mục đầu tư (Admin)",
        "tags": [
          "Portfolios"
        ]
      }
    },
    "/portfolios": {
      "get": {
        "description": "Trả về danh mục đầu tư mới nhất theo gói. Yêu cầu gói phù hợp.",
        "operationId": "PortfoliosController_findLatestByPlan",
        "parameters": [
          {
            "name": "plan",
            "required": true,
            "in": "query",
            "description": "Slug gói (vd: gold)",
            "schema": {
              "example": "gold",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh mục đầu tư"
          },
          "403": {
            "description": "Cần nâng cấp gói"
          },
          "404": {
            "description": "Không tìm thấy gói hoặc danh mục"
          }
        },
        "summary": "Danh mục đầu tư mới nhất",
        "tags": [
          "Portfolios"
        ]
      },
      "post": {
        "description": "Tạo danh mục đầu tư mới với cổ phiếu, thông tin, lý do, và tín hiệu",
        "operationId": "PortfoliosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePortfolioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Danh mục đầu tư đã được tạo"
          },
          "401": {
            "description": "Chưa đăng nhập"
          },
          "403": {
            "description": "Không có quyền"
          },
          "404": {
            "description": "Không tìm thấy gói"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo danh mục đầu tư",
        "tags": [
          "Portfolios"
        ]
      }
    },
    "/portfolios/{id}": {
      "patch": {
        "description": "Cập nhật danh mục. Dữ liệu lồng nhau sẽ được xóa và tạo lại.",
        "operationId": "PortfoliosController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePortfolioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Danh mục đầu tư đã được cập nhật"
          },
          "401": {
            "description": "Chưa đăng nhập"
          },
          "403": {
            "description": "Không có quyền"
          },
          "404": {
            "description": "Không tìm thấy danh mục đầu tư"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật danh mục đầu tư",
        "tags": [
          "Portfolios"
        ]
      },
      "delete": {
        "description": "Xóa danh mục đầu tư và tất cả dữ liệu liên quan. Chỉ admin.",
        "operationId": "PortfoliosController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh mục đầu tư đã được xóa"
          },
          "401": {
            "description": "Chưa đăng nhập"
          },
          "403": {
            "description": "Không có quyền"
          },
          "404": {
            "description": "Không tìm thấy danh mục đầu tư"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa danh mục đầu tư",
        "tags": [
          "Portfolios"
        ]
      }
    },
...
  },
  "info": {
    "title": "HotStock API Documentation",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email đăng nhập"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "Mật khẩu"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email đăng ký"
          },
          "username": {
            "type": "string",
            "example": "johndoe",
            "description": "Tên người dùng"
          },
          "fullName": {
            "type": "string",
            "example": "Nguyễn Văn A",
            "description": "Họ và tên"
          },
          "phoneNumber": {
            "type": "string",
            "example": "0912345678",
            "description": "Số điện thoại"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "Mật khẩu (tối thiểu 8 ký tự)"
          },
          "confirmPassword": {
            "type": "string",
            "example": "password123",
            "description": "Xác nhận mật khẩu"
          }
        },
        "required": [
          "email",
          "username",
          "password",
          "confirmPassword"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "Refresh token nhận được khi đăng nhập"
          }
        },
        "required": [
          "refresh_token"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "Mật khẩu hiện tại"
          },
          "newPassword": {
            "type": "string",
            "description": "Mật khẩu mới (tối thiểu 8 ký tự)"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email đã đăng ký"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email đã đăng ký"
          },
          "otp": {
            "type": "string",
            "example": "123456",
            "description": "Mã OTP 6 chữ số"
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "resetToken": {
            "type": "string",
            "description": "Reset token nhận được từ verify-otp"
          },
          "newPassword": {
            "type": "string",
            "description": "Mật khẩu mới (tối thiểu 8 ký tự)"
          }
        },
        "required": [
          "resetToken",
          "newPassword"
        ]
      },
...
      "PortfolioStockDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "example": "VNM"
          },
          "sector": {
            "type": "string",
            "example": "Năng lượng"
          },
          "purchaseDate": {
            "type": "string",
            "example": "2024-06-01T00:00:00.000Z"
          },
          "costBasis": {
            "type": "number",
            "example": 85000
          },
          "marketPrice": {
            "type": "number",
            "example": 92000
          },
          "quantity": {
            "type": "number",
            "example": 100
          },
          "note": {
            "type": "string",
            "example": "Mua thêm ở vùng hỗ trợ"
          }
        },
        "required": [
          "symbol",
          "purchaseDate",
          "costBasis",
          "marketPrice",
          "quantity"
        ]
      },
      "PortfolioInformationDto": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "example": "2024-06"
          },
          "vnindexReturn": {
            "type": "number",
            "example": 5.2
          },
          "recommendReturn": {
            "type": "number",
            "example": 8.7
          }
        },
        "required": [
          "month",
          "vnindexReturn",
          "recommendReturn"
        ]
      },
      "PortfolioReasonDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "buy",
            "enum": [
              "buy",
              "sell"
            ]
          },
          "symbol": {
            "type": "string",
            "example": "VNM"
          },
          "content": {
            "type": "string",
            "example": "Kết quả kinh doanh Q2 tốt hơn kỳ vọng"
          }
        },
        "required": [
          "type",
          "symbol",
          "content"
        ]
      },
      "PortfolioSignalDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "example": "VNM"
          },
          "signalType": {
            "type": "string",
            "example": "breakout"
          },
          "description": {
            "type": "string",
            "example": "Breakout khỏi vùng kháng cự 90,000"
          },
          "targetPrice": {
            "type": "number",
            "example": 95000
          },
          "stopLoss": {
            "type": "number",
            "example": 87000
          }
        },
        "required": [
          "symbol",
          "signalType",
          "description"
        ]
      },
      "CreatePortfolioDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "number",
            "example": 1,
            "description": "ID gói"
          },
          "publishedAt": {
            "type": "string",
            "example": "2024-06-15T00:00:00.000Z",
            "description": "Ngày công bố"
          },
          "stocks": {
            "description": "Danh sách cổ phiếu",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioStockDto"
            }
          },
          "information": {
            "description": "Thông tin hiệu suất",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioInformationDto"
            }
          },
          "reasons": {
            "description": "Lý do mua/bán",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioReasonDto"
            }
          },
          "signals": {
            "description": "Tín hiệu giao dịch",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioSignalDto"
            }
          }
        },
        "required": [
          "planId",
          "publishedAt",
          "stocks",
          "information",
          "reasons",
          "signals"
        ]
      },
      "UpdatePortfolioDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "number",
            "example": 1,
            "description": "ID gói"
          },
          "publishedAt": {
            "type": "string",
            "example": "2024-06-15T00:00:00.000Z",
            "description": "Ngày công bố"
          },
          "stocks": {
            "description": "Danh sách cổ phiếu",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioStockDto"
            }
          },
          "information": {
            "description": "Thông tin hiệu suất",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioInformationDto"
            }
          },
          "reasons": {
            "description": "Lý do mua/bán",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioReasonDto"
            }
          },
          "signals": {
            "description": "Tín hiệu giao dịch",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioSignalDto"
            }
          }
        }
      },
...
    }
  }
}